+++ canonicalwebteam/store_base/data/licenses.json
