+++ canonicalwebteam/store_base/sample_blueprint/views.py
@@ -1,11 +1,10 @@
 from flask import Blueprint
 
-test_bp = Blueprint(
+sample_bp = Blueprint(
     "test_store_bp",
     __name__,
-
 )
 
-@test_bp.route("/test_store_bp")
-def test_bp_route():
+@sample_bp.route("/test_store_bp")
+def sample_route():
     return "This is a test blueprint on the storebase"
