+++ canonicalwebteam/store_base/utils/helpers.py
